# doc/swagger-doc.yaml
openapi: 3.0.0
info:
  title: Generic Xumm API
  description: |
   This API is dedicated to interact with Xumm and will be improved over time.
   This API is a full mirror of the [Xumm API](https://xumm.readme.io/reference/about) by [@XRPLLabs](https://twitter.com/XRPLLabs).<br>
   The prupose of this mirror API is to provide easy access to Xumm for, as example, Website developers, Alexa Skill developers, ShopPlugin developers
   and others, without the need for them to implement their own backend.
   
   The API supports all methods the Xumm API offers and also offers additional capabilities to make it easier to interact with Xumm.
   
   <b>This API only describes the additional parameters the Xumm API does not offer.</b><br>
   To check how to interact with Xumm and which fields to use, please refer to: [Xumm API](https://xumm.readme.io/reference/about)



   <b>To use this API you have to be whitlisted.</b> Please contact [@nixerFFM](https://twitter.com/nixerFFM) if you are interested to use this backend API.
  version: 1.4.0
paths:
  /api/v1/platform/payload:
    post:
      summary: Submits a payload to Xumm
      description: Submits a payload to Xumm and returns information about the submit result. Payload field specification can be found here [Xumm API Reference](https://xumm.readme.io/reference/post-payload)
                    <br>Additional parameters below
      tags:
        - Xumm API
      produces:
        - application/json
      parameters:
        - name: frontendId
          description: Unique ID to identify a user of your application. If included, push notifications to users are possible.
          in: body
          required: false
          schema:
            type: string
        - name: pushDisabled
          description: enable or disable push notifications for a specific user (only required if frontendId is available)
          in: body
          required: false
          schema:
            type: boolean
        - name: web
          description: specify where the request is coming from. web=true -> Desktop, web=false -> App. This is an important setting and determines which return_url is being set (app or web)
          in: body
          required: false
          schema:
            type: boolean
      responses:
        200:
          description: Returns a payload submit response. Please refer to (https://xumm.readme.io/reference/post-payload)
          content:
            application/json:
              schema:
                type: object
        500:
          description: Error. Is returned when no origin is set in the request or an exceptions occurs when handling your request. Please check your request parameters.
          schema:
            type: string
  /api/v1/platform/payload/:payload_uuid:
    get:
      summary: Get the payload info
      description: Returns the payload information of a specific payloadId. Specification about the return fields can be found here [Xumm API Reference](https://xumm.readme.io/reference/get-payload)
      tags:
        - Xumm API
      produces:
        - application/json
      parameters:
        - name: payload_uuid
          description: id of a specific payload
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Returns information about the payloadId. Please refer to (https://xumm.readme.io/reference/get-payload)
          content:
            application/json:
              schema:
                type: object
        500:
          description: Error. Is returned when no origin is set in the request or an exceptions occurs when handling your request. Please check your request parameters.
          schema:
            type: string
    delete:
      summary: Cancel a payload
      description: Cancel a payload, so a user cannot open it anymore. Specification about the return fields can be found here [Xumm API Reference](https://xumm.readme.io/reference/cancel-payload)
      tags:
        - Xumm API
      produces:
        - application/json
      parameters:
        - name: id
          description: id of a specific payload
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Returns the status of the delete. Please refer to (https://xumm.readme.io/reference/cancel-payload)
          content:
            application/json:
              schema:
                type: object
        500:
          description: Error. Is returned when no origin is set in the request or an exceptions occurs when handling your request. Please check your request parameters.
          schema:
            type: string
  /api/v1/check/payment/:payload_uuid:
    get:
      summary: Returns the validation status of a payment payload
      description: Validates a payment payload. Does check the payment payload properties AND does check the transaction on the XRP Ledger. Checks the transaction on the XRPL live net AND test net!
      tags:
        - Check API
      produces:
        - application/json
      parameters:
        - name: payload_uuid
          description: id of the payment payload to check
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Returns the result of the payment payload validation. 
          content:
            application/json:
              schema:
                type: object
                properties:
                    success:
                      description: true, if transactions was found and validated on XRPL live net OR test net
                      type: boolean
                      example: true
                    testnet:
                      description: true, if transaction was found on test net. false, if transaction was found on live net
                      type: boolean
                      example: false
        500:
          description: Error. Is returned when no origin is set in the request or an exceptions occurs when handling your request. Please check your request parameters.
          schema:
            type: string
  /api/v1/check/payment/:frontendUserId/:payload_uuid:
    get:
      summary: Returns the validation status of a payment payload signed by a specific user
      description: Validates a payment payload. Checks if a payment payload can be matched to a user (signed by the user). Checks also the payload properties AND does check the transaction on the XRP Ledger. Checks the transaction on the XRPL live net AND test net!
      tags:
        - Check API
      produces:
        - application/json
      parameters:
        - name: frontendUserId
          description: id of the user you want to check the payment payload for
          in: path
          required: true
          schema:
            type: string
        - name: payload_uuid
          description: id of the payment payload to check
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Returns the result of the payment payload validation. 
          content:
            application/json:
              schema:
                type: object
                properties:
                    success:
                      description: true, if transactions was found and validated on XRPL live net OR test net
                      type: boolean
                      example: true
                    testnet:
                      description: true, if transaction was found on test net. false, if transaction was found on live net
                      type: boolean
                      example: true
        500:
          description: Error. Is returned when no origin is set in the request or an exceptions occurs when handling your request. Please check your request parameters.
          schema:
            type: string
  /api/v1/check/timed/payment/:payload_uuid:
    get:
      summary: Returns the validation status of a payment payload for a given time (currently 24h)
      description: Validates a payment payload agains an expiration time (currently 24h). Does check the payload properties AND does check the transaction on the XRP Ledger. Checks the transaction on the XRPL live net AND test net!
      tags:
        - Check API
      produces:
        - application/json
      parameters:
        - name: payload_uuid
          description: id of the payment payload to check
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Returns the result of the payment payload validation. 
          content:
            application/json:
              schema:
                type: object
                properties:
                    success:
                      description: true, if payment payload is within timeframe (currently 24h) and transactions was found and validated on XRPL live net OR test net
                      type: boolean
                      example: false
                    testnet:
                      description: true, if transaction was found on test net. false, if transaction was found on live net
                      type: boolean
                      example: false
        500:
          description: Error. Is returned when no origin is set in the request or an exceptions occurs when handling your request. Please check your request parameters.
          schema:
            type: string
  /api/v1/check/timed/payment/:frontendUserId/:payload_uuid:
    get:
      summary: Returns the validation status of a payment payload signed by the specified user and for a given time
      description: Validates a payment payload agains an expiration time (currently 24h). Checks if a payload can be matched to a user (signed by this user). Checks also the payload properties AND does check the transaction on the XRP Ledger. Checks the transaction on the XRPL live net AND test net!
      tags:
        - Check API
      produces:
        - application/json
      parameters:
        - name: frontendUserId
          description: id of the user you want to check the payment payload for
          in: path
          required: true
          schema:
            type: string
        - name: payload_uuid
          description: id of the payment payload to check
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Returns the result of the payment payload validation. 
          content:
            application/json:
              schema:
                type: object
                properties:
                    success:
                      description: true, if payload is within timeframe (currently 24h) and if transactions was found and validated on XRPL live net OR test net
                      type: boolean
                      example: false
                    testnet:
                      description: true, if transaction was found on test net. false, if transaction was found on live net
                      type: boolean
                      example: true
        500:
          description: Error. Is returned when no origin is set in the request or an exceptions occurs when handling your request. Please check your request parameters.
          schema:
            type: string
  /api/v1/check/sigin/:payload_uuid:
    get:
      summary: Returns the validation status of a sigin request
      description: Validates a sigin payload. Does check the signin payload properties. No On Ledger validation for SignIn requests.
      tags:
        - Check API
      produces:
        - application/json
      parameters:
        - name: payload_uuid
          description: id of the sigin payload to check
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Returns the result of the signin payload validation. 
          content:
            application/json:
              schema:
                type: object
                properties:
                    success:
                      description: true, if signin request is valid
                      type: boolean
                      example: true
        500:
          description: Error. Is returned when no origin is set in the request or an exceptions occurs when handling your request. Please check your request parameters.
          schema:
            type: string
  /api/v1/check/sigin/:frontendUserId/:payload_uuid:
    get:
      summary: Returns the validation status of a signin payload signed by the specified user
      description: Validates a signin payload. Checks if a signin payload can be matched to a user (signed by this user). Checks also the singin payload properties AND if it was signed successfully.
      tags:
        - Check API
      produces:
        - application/json
      parameters:
        - name: frontendUserId
          description: id of the user you want to check the signin payload for
          in: path
          required: true
          schema:
            type: string
        - name: payload_uuid
          description: id of the signin payload to check
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Returns the result of the signin payload validation. 
          content:
            application/json:
              schema:
                type: object
                properties:
                    success:
                      description: true, if signin payload was validated and signed
                      type: boolean
                      example: false
        500:
          description: Error. Is returned when no origin is set in the request or an exceptions occurs when handling your request. Please check your request parameters.
          schema:
            type: string
  /api/v1/special/signinToValidate/:signinPayload_uuid:
    get:
      summary: Returns the validation status of a payment found for the XRPL account which signed the SignIn payload
      description: | 
        This method matched the XRPL account used to sign the SignIn payload to a payment payload the same XRPL account has signed.<br>
        If the valid payment payload is not older than 24h, this method returns 'success'. (same as '/api/v1/check/timed/payment/{payloadId}' but with matching of the XRPL account from the SignIn request).<br>
        Can be used to allow users accessing their purchased goods (statistics, charts, ...) on different environments/machines.
      tags:
        - Special API
      produces:
        - application/json
      parameters:
        - name: signinPayload_uuid
          description: id of a users signin request
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Returns the result of the signin payload validation. 
          content:
            application/json:
              schema:
                type: object
                properties:
                    success:
                      description: true, if payload is within timeframe (currently 24h) and if transactions was found and validated on XRPL live net OR test net
                      type: boolean
                      example: false
                    testnet:
                      description: true, if transaction was found on test net. false, if transaction was found on live net
                      type: boolean
                      example: true
        500:
          description: Error. Is returned when no origin is set in the request or an exceptions occurs when handling your request. Please check your request parameters.
          schema:
            type: string
  